{"schema":"decodex/commit/1","summary":"Extract orchestrator execution failure modules","authority":"XY-1115"}
