+++ apps/decodex/src/orchestrator/execution_failure/terminal_writeback.rs
@@ -1,0 +1,319 @@
+use super::{
+	HarnessOutcomeKind, IssueRunPlan, IssueTracker, PreparedTerminalFailureWriteback, Report,
+	Result, RetainedPartialProgress, TerminalFailureEventRecordStatus, TerminalFailureLifecycle,
+	TerminalFailureOutcome, TerminalFailureWritebackRuntime, WorkflowDocument,
+	ensure_automation_activity_label, eyre, format_terminal_failure_comment,
+	record_harness_outcome_best_effort, records, terminal_failure_comment_details,
+	terminal_failure_lifecycle_event, terminal_failure_pr_url, terminal_failure_recovery_gate,
+	tracker,
+};
+
+pub(in crate::orchestrator) fn apply_terminal_failure_writeback<T>(
+	tracker: &T,
+	runtime: TerminalFailureWritebackRuntime<'_>,
+	workflow: &WorkflowDocument,
+	issue_run: &IssueRunPlan,
+	worktree_path: &str,
+	manual_attention_requested: bool,
+	error: &Report,
+) -> Result<TerminalFailureOutcome>
+where
+	T: IssueTracker,
+{
+	let writeback = prepare_terminal_failure_writeback(
+		tracker,
+		runtime,
+		workflow,
+		issue_run,
+		worktree_path,
+		manual_attention_requested,
+		error,
+	)?;
+	let event_status =
+		record_terminal_failure_writeback_event(tracker, runtime, issue_run, &writeback)?;
+
+	if event_status == TerminalFailureEventRecordStatus::Duplicate {
+		return Ok(terminal_failure_outcome(&writeback));
+	}
+
+	let writeback_result =
+		apply_terminal_failure_tracker_writeback(tracker, runtime, issue_run, &writeback);
+
+	if let Err(error) = writeback_result {
+		forget_terminal_failure_writeback_event(runtime, event_status, &writeback)?;
+
+		return Err(error);
+	}
+	if let Some(state_store) = runtime.state_store {
+		let outcome = if writeback.projection.record.event_type == "needs_attention" {
+			HarnessOutcomeKind::ManualAttention
+		} else {
+			HarnessOutcomeKind::TerminalFailure
+		};
+
+		record_harness_outcome_best_effort(
+			state_store,
+			runtime.service_id,
+			issue_run,
+			outcome,
+			Some(writeback.error_class),
+			None,
+			writeback.projection.record.pr_url.as_deref(),
+		);
+	}
+
+	Ok(terminal_failure_outcome(&writeback))
+}
+
+fn prepare_terminal_failure_writeback<T>(
+	tracker: &T,
+	runtime: TerminalFailureWritebackRuntime<'_>,
+	workflow: &WorkflowDocument,
+	issue_run: &IssueRunPlan,
+	worktree_path: &str,
+	manual_attention_requested: bool,
+	error: &Report,
+) -> Result<PreparedTerminalFailureWriteback>
+where
+	T: IssueTracker,
+{
+	let tracker_policy = workflow.frontmatter().tracker();
+	let needs_attention_label = tracker_policy.needs_attention_label();
+	let needs_attention_label_id = tracker::issue_team_label_id_with_server_confirmation(
+		tracker,
+		&issue_run.issue,
+		needs_attention_label,
+	)?;
+	let failure_state_name = tracker_policy.failure_state();
+	let failure_state_is_startable =
+		tracker_policy.startable_states().iter().any(|state| state == failure_state_name);
+	let retry_guarded_by_state = needs_attention_label_id.is_none() && failure_state_is_startable;
+	let terminal_failure_state_name = if retry_guarded_by_state {
+		tracker_policy.in_progress_state()
+	} else {
+		failure_state_name
+	};
+	let failure_state_id =
+		issue_run.issue.state_id_for_name(terminal_failure_state_name).ok_or_else(|| {
+			eyre::eyre!(
+				"State `{}` was not found for issue `{}`.",
+				terminal_failure_state_name,
+				issue_run.issue.identifier
+			)
+		})?;
+	let recovery_gate = terminal_failure_recovery_gate(
+		needs_attention_label,
+		needs_attention_label_id.is_some(),
+		retry_guarded_by_state,
+		tracker_policy.in_progress_state(),
+	);
+	let (error_class, next_action) =
+		terminal_failure_comment_details(manual_attention_requested, error, &recovery_gate);
+	let pr_url = terminal_failure_pr_url(error);
+	let retained_source_error_class = error
+		.downcast_ref::<RetainedPartialProgress>()
+		.and_then(|partial_progress| partial_progress.source_error_class.as_deref());
+	let comment = format_terminal_failure_comment(
+		&issue_run.run_id,
+		issue_run.attempt_number,
+		worktree_path.to_owned(),
+		&issue_run.worktree.branch_name,
+		pr_url,
+		error_class,
+		&next_action,
+	);
+	let event = terminal_failure_lifecycle_event(
+		runtime.service_id,
+		issue_run,
+		TerminalFailureLifecycle {
+			error_class,
+			next_action: &next_action,
+			pr_url,
+			target_state: terminal_failure_state_name,
+			worktree_path,
+			manual_attention_requested,
+			retained_source_error_class,
+		},
+	);
+	let projection = tracker::prepare_linear_execution_event_comment(
+		&comment,
+		&event,
+		runtime.privacy_classifier,
+	)?;
+
+	Ok(PreparedTerminalFailureWriteback {
+		failure_state_id: failure_state_id.to_owned(),
+		needs_attention_label: needs_attention_label.to_owned(),
+		needs_attention_label_id,
+		terminal_failure_state_name: terminal_failure_state_name.to_owned(),
+		projection,
+		error_class,
+		retry_guarded_by_state,
+	})
+}
+
+fn record_terminal_failure_writeback_event<T>(
+	tracker: &T,
+	runtime: TerminalFailureWritebackRuntime<'_>,
+	issue_run: &IssueRunPlan,
+	writeback: &PreparedTerminalFailureWriteback,
+) -> Result<TerminalFailureEventRecordStatus>
+where
+	T: IssueTracker,
+{
+	let event_status = if let Some(state_store) = runtime.state_store {
+		if !state_store.record_linear_execution_event(&writeback.projection.record)? {
+			return Ok(TerminalFailureEventRecordStatus::Duplicate);
+		}
+
+		TerminalFailureEventRecordStatus::Recorded
+	} else {
+		TerminalFailureEventRecordStatus::NoLocalStore
+	};
+
+	if remote_terminal_failure_writeback_exists(
+		tracker,
+		runtime,
+		issue_run,
+		writeback,
+		event_status,
+	)? {
+		return Ok(TerminalFailureEventRecordStatus::Duplicate);
+	}
+
+	Ok(event_status)
+}
+
+fn remote_terminal_failure_writeback_exists<T>(
+	tracker: &T,
+	runtime: TerminalFailureWritebackRuntime<'_>,
+	issue_run: &IssueRunPlan,
+	writeback: &PreparedTerminalFailureWriteback,
+	event_status: TerminalFailureEventRecordStatus,
+) -> Result<bool>
+where
+	T: IssueTracker,
+{
+	let comments = match tracker.list_comments(&issue_run.issue.id) {
+		Ok(comments) => comments,
+		Err(error) => {
+			forget_terminal_failure_writeback_event(runtime, event_status, writeback)?;
+
+			return Err(error);
+		},
+	};
+
+	if !records::has_linear_execution_event_record(
+		&comments,
+		&writeback.projection.record.service_id,
+		&writeback.projection.record.issue_id,
+		&writeback.projection.record.idempotency_key,
+	) {
+		return Ok(false);
+	}
+
+	tracing::debug!(
+		service_id = writeback.projection.record.service_id,
+		issue_id = issue_run.issue.id,
+		issue = issue_run.issue.identifier,
+		run_id = issue_run.run_id,
+		attempt = issue_run.attempt_number,
+		event_type = writeback.projection.record.event_type,
+		"Skipping terminal failure writeback already present in remote Linear ledger."
+	);
+
+	Ok(true)
+}
+
+fn apply_terminal_failure_tracker_writeback<T>(
+	tracker: &T,
+	runtime: TerminalFailureWritebackRuntime<'_>,
+	issue_run: &IssueRunPlan,
+	writeback: &PreparedTerminalFailureWriteback,
+) -> Result<()>
+where
+	T: IssueTracker,
+{
+	tracker.update_issue_state(&issue_run.issue.id, &writeback.failure_state_id)?;
+
+	apply_needs_attention_label(
+		tracker,
+		issue_run,
+		runtime.service_id,
+		&writeback.needs_attention_label,
+		writeback.needs_attention_label_id.clone(),
+		&writeback.terminal_failure_state_name,
+	)?;
+
+	if runtime.state_store.is_some() {
+		tracker::create_prepared_linear_execution_event_comment_without_remote_scan(
+			tracker,
+			&issue_run.issue.id,
+			&writeback.projection,
+		)?;
+	} else {
+		tracker::create_prepared_linear_execution_event_comment(
+			tracker,
+			&issue_run.issue.id,
+			&writeback.projection,
+		)?;
+	}
+
+	Ok(())
+}
+
+fn forget_terminal_failure_writeback_event(
+	runtime: TerminalFailureWritebackRuntime<'_>,
+	event_status: TerminalFailureEventRecordStatus,
+	writeback: &PreparedTerminalFailureWriteback,
+) -> Result<()> {
+	if event_status == TerminalFailureEventRecordStatus::Recorded
+		&& let Some(state_store) = runtime.state_store
+	{
+		state_store.forget_linear_execution_event(&writeback.projection.record.idempotency_key)?;
+	}
+
+	Ok(())
+}
+
+fn terminal_failure_outcome(
+	writeback: &PreparedTerminalFailureWriteback,
+) -> TerminalFailureOutcome {
+	TerminalFailureOutcome {
+		error_class: writeback.error_class,
+		retry_guarded_by_state: writeback.retry_guarded_by_state,
+	}
+}
+
+fn apply_needs_attention_label<T>(
+	tracker: &T,
+	issue_run: &IssueRunPlan,
+	service_id: &str,
+	needs_attention_label: &str,
+	needs_attention_label_id: Option<String>,
+	terminal_failure_state_name: &str,
+) -> Result<bool>
+where
+	T: IssueTracker,
+{
+	if let Some(label_id) = needs_attention_label_id.as_deref() {
+		if !tracker::issue_has_label_with_server_confirmation(
+			tracker,
+			&issue_run.issue,
+			needs_attention_label,
+		)? {
+			tracker.add_issue_labels(&issue_run.issue.id, &[label_id.to_owned()])?;
+		}
+	} else {
+		tracing::warn!(
+			label = needs_attention_label,
+			issue = issue_run.issue.identifier,
+			guard_state = terminal_failure_state_name,
+			"Needs-attention label was not found in the issue team; using a non-startable state guard when needed."
+		);
+	}
+
+	ensure_automation_activity_label(tracker, &issue_run.issue, service_id, false)?;
+
+	Ok(needs_attention_label_id.is_some())
+}
